Consolidating Source Code
+ Moving Verbs back into the main package, to reduce code duplication
while maintaining encapsulation
