+++ src/main/java/tech/sirwellington/alchemy/http/BaseVerb.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package tech.sirwellington.alchemy.http.verb;
+package tech.sirwellington.alchemy.http;
 
 import com.google.common.base.Charsets;
 import com.google.common.base.Strings;
@@ -37,15 +37,13 @@ import org.slf4j.LoggerFactory;
 import tech.sirwellington.alchemy.annotations.access.Internal;
 import tech.sirwellington.alchemy.annotations.arguments.NonNull;
 import tech.sirwellington.alchemy.arguments.FailedAssertionException;
-import tech.sirwellington.alchemy.http.HttpRequest;
-import tech.sirwellington.alchemy.http.HttpResponse;
 import tech.sirwellington.alchemy.http.exceptions.AlchemyHttpException;
 import tech.sirwellington.alchemy.http.exceptions.JsonException;
 import tech.sirwellington.alchemy.http.exceptions.OperationFailedException;
 
 import static tech.sirwellington.alchemy.arguments.Arguments.checkThat;
 import static tech.sirwellington.alchemy.arguments.Assertions.notNull;
-import static tech.sirwellington.alchemy.http.verb.VerbAssertions.validContentType;
+import static tech.sirwellington.alchemy.http.HttpAssertions.validContentType;
 
 /**
  *
@@ -54,42 +52,43 @@ import static tech.sirwellington.alchemy.http.verb.VerbAssertions.validContentTy
 @Internal
 final class BaseVerb implements HttpVerb
 {
-
+    
     private final static Logger LOG = LoggerFactory.getLogger(BaseVerb.class);
-
+    
     private final Gson gson = new GsonBuilder()
+            .setDateFormat(Constants.DATE_FORMAT)
             .create();
-
+    
     private final AlchemyRequestMapper requestMapper;
-
+    
     BaseVerb(AlchemyRequestMapper requestMapper)
     {
         checkThat(requestMapper).is(notNull());
-
+        
         this.requestMapper = requestMapper;
     }
-
+    
     static BaseVerb using(AlchemyRequestMapper requestMapper)
     {
         return new BaseVerb(requestMapper);
     }
-
+    
     @Override
     public HttpResponse execute(HttpClient apacheHttpClient, HttpRequest request) throws AlchemyHttpException
     {
         checkThat(apacheHttpClient, request)
                 .usingMessage("null arguments")
                 .are(notNull());
-
+        
         HttpUriRequest apacheRequest = requestMapper.convertToApacheRequest(request);
-
+        
         checkThat(apacheRequest)
                 .throwing(ex -> new AlchemyHttpException("Could not map HttpRequest: " + request))
                 .is(notNull());
-
+        
         request.getRequestHeaders()
                 .forEach(apacheRequest::addHeader);
-
+        
         org.apache.http.HttpResponse apacheResponse;
         try
         {
@@ -100,11 +99,11 @@ final class BaseVerb implements HttpVerb
             LOG.error("Failed to execute GET Request on {}", request, ex);
             throw new AlchemyHttpException(ex);
         }
-
+        
         checkThat(apacheResponse)
                 .throwing(ex -> new AlchemyHttpException(request, "Missing Apache Client Response"))
                 .is(notNull());
-
+        
         JsonElement json;
         try
         {
@@ -120,31 +119,31 @@ final class BaseVerb implements HttpVerb
             LOG.error("Could not parse Response from Request {}", request, ex);
             throw new OperationFailedException(request, ex);
         }
-
+        
         HttpResponse response = HttpResponse.Builder.newInstance()
                 .withResponseBody(json)
                 .withStatusCode(apacheResponse.getStatusLine().getStatusCode())
                 .withResponseHeaders(extractHeadersFrom(apacheResponse))
                 .usingGson(gson)
                 .build();
-
+        
         return response;
     }
-
+    
     private JsonElement extractJsonFromResponse(@NonNull org.apache.http.HttpResponse apacheResponse) throws JsonException, JsonParseException
     {
         if (apacheResponse.getEntity() == null)
         {
             return JsonNull.INSTANCE;
         }
-
+        
         HttpEntity entity = apacheResponse.getEntity();
-
+        
         String contentType = entity.getContentType().getValue();
         checkThat(contentType)
                 .throwing(JsonException.class)
                 .is(validContentType());
-
+        
         String responseString = null;
         try (final InputStream istream = entity.getContent())
         {
@@ -160,7 +159,7 @@ final class BaseVerb implements HttpVerb
             LOG.error("Failed to read entity from request", ex);
             throw new AlchemyHttpException("Failed to read response from server", ex);
         }
-
+        
         if (Strings.isNullOrEmpty(responseString))
         {
             return JsonNull.INSTANCE;
@@ -170,17 +169,17 @@ final class BaseVerb implements HttpVerb
             return gson.fromJson(responseString, JsonElement.class);
         }
     }
-
+    
     private Map<String, String> extractHeadersFrom(org.apache.http.HttpResponse apacheResponse)
     {
         if (apacheResponse.getAllHeaders() == null)
         {
             return Collections.emptyMap();
         }
-
+        
         return Arrays.asList(apacheResponse.getAllHeaders())
                 .stream()
                 .collect(Collectors.toMap(Header::getName, Header::getValue));
     }
-
+    
 }
