+++ src/test/java/tech/sirwellington/alchemy/http/CommonAssertions.java
@@ -29,7 +29,6 @@ import org.apache.http.message.BasicHttpResponse;
 import org.mockito.ArgumentCaptor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import tech.sirwellington.alchemy.http.verb.HttpVerb;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
