--- src/test/java/tech/sirwellington/alchemy/http/verb/VerbAssertionsTest.java
@@ -1,81 +1,0 @@
-/*
- * Copyright 2015 SirWellington Tech.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tech.sirwellington.alchemy.http.verb;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.mockito.Mock;
-import tech.sirwellington.alchemy.arguments.AlchemyAssertion;
-import tech.sirwellington.alchemy.arguments.FailedAssertionException;
-import tech.sirwellington.alchemy.generator.AlchemyGenerator;
-import tech.sirwellington.alchemy.generator.StringGenerators;
-
-import static org.junit.Assert.*;
-import static org.hamcrest.Matchers.*;
-import static org.mockito.Mockito.*;
-import static tech.sirwellington.alchemy.generator.AlchemyGenerator.one;
-import static tech.sirwellington.alchemy.generator.StringGenerators.alphabeticString;
-import static tech.sirwellington.alchemy.generator.StringGenerators.hexadecimalString;
-import static tech.sirwellington.alchemy.test.junit.ThrowableAssertion.assertThrows;
-
-/**
- *
- * @author SirWellington
- */
-@RunWith(MockitoJUnitRunner.class)
-public class VerbAssertionsTest
-{
-
-    @Before
-    public void setUp()
-    {
-    }
-
-    @Test
-    public void testValidContentType()
-    {
-        AlchemyAssertion<String> instance = VerbAssertions.validContentType();
-        assertThat(instance, notNullValue());
-
-        AlchemyGenerator<String> validTypes = StringGenerators.stringsFromFixedList("application/json", "text/plain");
-
-        String contentType = one(validTypes);
-
-        instance.check(contentType);
-        instance.check(contentType + one(alphabeticString()));
-    }
-
-    @Test
-    public void testValidContentTypeEdgeCases()
-    {
-        AlchemyAssertion<String> instance = VerbAssertions.validContentType();
-
-        //Edge cases
-        assertThrows(() -> instance.check(null))
-                .isInstanceOf(FailedAssertionException.class);
-
-        assertThrows(() -> instance.check(""))
-                .isInstanceOf(FailedAssertionException.class);
-
-        assertThrows(() -> instance.check(one(alphabeticString())))
-                .isInstanceOf(FailedAssertionException.class);
-
-        assertThrows(() -> instance.check(one(hexadecimalString(10))))
-                .isInstanceOf(FailedAssertionException.class);
-    }
-}
